requirements.txt
amqp==5.0.2
celery==5.0.4
beautifulsoup4==4.9.3
Django==3.1.4
django-celery-beat==2.1.0
django-celery-results==2.0.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
hyperopt==0.2.5
joblib==1.0.0
numpy==1.19.5
pandas==1.2.0
pytest==6.2.1
pytest-django==4.1.0
python-crontab==2.5.1
psycopg2==2.7.5
ray[tune]
requests==2.25.1
scikit-learn==0.24.0
tensorflow==2.4.0

